rockchip.wks: use uuid for /boot during fstab-update
Since the recent patch to switch to UUIDs [0aa5e600: "use uuid
instead of hard-coding root device"] wic fstab-update is not able
to get the correct value for the used device anymore and falls to
the default 'sda'. Thus wrong /dev/sda entries are generated in fstab.

For partitions that should be updated automatically this can be avoided
by either generate entries using uuid or label.

[NOTE: I rearranged the order of the arguments so they line up]

Signed-off-by: MarkusVolk <f_l_k@t-online.de>